requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
alembic==1.8.1
    # via flask-migrate
click==8.1.3
    # via flask
colorama==0.4.5
    # via click
dnspython==2.2.1
    # via -r requirements.in
flask==2.2.2
    # via
    #   -r requirements.in
    #   flask-migrate
    #   flask-sqlalchemy
flask-migrate==3.1.0
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements.in
    #   flask-migrate
greenlet==1.1.2
    # via sqlalchemy
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
mako==1.2.1
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
    #   werkzeug
mistune==2.0.4
    # via -r requirements.in
mysqlclient==2.1.1
    # via -r requirements.in
paste==3.5.2
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
six==1.16.0
    # via paste
sqlalchemy==1.4.40
    # via
    #   alembic
    #   flask-sqlalchemy
ua-parser==0.16.0
    # via user-agents
user-agents==2.2.0
    # via -r requirements.in
waitress==2.1.2
    # via -r requirements.in
werkzeug==2.2.2
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
